Move Log UIi component out of gen-widgets (since it's static)
