+++ example/web-ui/src/Log.tsx
@@ -1,0 +1,42 @@
+import { useState, useEffect } from "react";
+import { addTopicCallback } from "./mqclient";
+
+
+interface LogMessage {
+  text: string;
+  value: number;
+  count: number;
+}
+export function Log({}) {
+  const [LogState, setLogState] = useState([{id: 0, text: ""}]);
+
+  function appendToLog(newLogMessage: LogMessage) {
+    setLogState(prevLogState => {
+      if (newLogMessage.count > prevLogState[prevLogState.length - 1].id)
+      {
+        let newState = prevLogState.slice(Math.max(prevLogState.length - 10, 0));
+        newState.push({
+          id: newLogMessage.count,
+          text: newLogMessage.count + "	" + newLogMessage.text + newLogMessage.value
+        });
+        return newState;
+      } else {
+        // Enforce log message ID's always increase to avoid duplicate keys.
+        return prevLogState;
+      }
+    });
+  }
+  useEffect( () => {
+    addTopicCallback("Log", appendToLog);
+  }, []);
+
+  const messages = LogState.map((message) => 
+    <p key={message.id}>{message.text}</p>)
+
+  return (
+    <div className="widget">
+      <h4>Log</h4>
+      {messages}
+    </div>
+  )
+}
